chore: move util function to utils folder
