+++ src/constants/index.ts
@@ -1,6 +1,5 @@
 export * from './locales'
 export * from './routes'
-export * from './dropdowns'
 export * from './applications'
 export * from './release'
 export * from './navigations'
